auth/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
idna==3.4
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.3
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
python-decouple==3.6
pytz==2022.6
redis==4.3.4
requests==2.28.1
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
tomli==2.0.1
urllib3==1.26.12
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.5.1
